@@ -11,31 +11,30 @@
     "postinstall": "nuxt prepare"
   },
   "devDependencies": {
-    "@lehoczky/eslint-config-vue": "^0.0.10",
+    "@lehoczky/eslint-config-vue": "^0.0.14",
     "@types/common-tags": "^1.8.1",
-    "@vueuse/core": "^9.5.0",
-    "@vueuse/nuxt": "^9.5.0",
-    "eslint": "^8.27.0",
-    "eslint-define-config": "^1.12.0",
-    "nuxt": "3.0.0-rc.11",
-    "postcss": "8.4.19",
+    "@vueuse/core": "^9.13.0",
+    "@vueuse/nuxt": "^9.13.0",
+    "eslint": "^8.38.0",
+    "eslint-define-config": "^1.17.0",
+    "nuxt": "3.4.0",
+    "postcss": "8.4.21",
     "postcss-html": "^1.5.0",
-    "prettier": "^2.7.1",
-    "prettier-plugin-tailwindcss": "^0.1.13",
-    "stylelint": "^14.14.1",
-    "stylelint-config-prettier": "^9.0.4",
-    "stylelint-config-recommended-vue": "^1.4.0",
-    "stylelint-config-standard": "^29.0.0",
+    "prettier": "^2.8.7",
+    "prettier-plugin-tailwindcss": "^0.2.7",
+    "stylelint": "^15.4.0",
+    "stylelint-config-recommended-vue": "1.4.0",
+    "stylelint-config-standard": "^31.0.0",
     "stylelint-config-tailwindcss": "^0.0.7",
-    "tailwindcss": "^3.2.4",
-    "typescript": "^4.8.4"
+    "tailwindcss": "^3.3.1",
+    "typescript": "^5.0.4"
   },
   "dependencies": {
-    "@headlessui/vue": "^1.7.4",
-    "@heroicons/vue": "^2.0.13",
+    "@headlessui/vue": "^1.7.12",
+    "@heroicons/vue": "^2.0.17",
     "@highlightjs/vue-plugin": "^2.1.0",
     "common-tags": "^1.8.2",
-    "highlight.js": "^11.6.0"
+    "highlight.js": "^11.7.0"
   },
   "pnpm": {
     "peerDependencyRules": {
